+++ ReleaseNotes/release_notes.md
@@ -17,6 +17,7 @@
    1. soem adjustment for `CleanUp` machinery 
    1. improvement for the output for `add_chopping_response` 
    1. do not add `sumw2`/`asymttotic` flags for fitting of `ROOT.RooDataHist` 
+   1. allow to deal with the histograms in `make_toys` and `make_toys2` 
  
 ## Backward incompatible: